apps/simulation/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
numpy>=1.26.0
scipy>=1.12.0
httpx>=0.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0
flake8>=7.0.0
pytest>=8.0.0

# Multi-Agent Orchestration
camel-oasis>=0.1.2
camel-ai>=0.1.5

# Knowledge Construction
graphrag>=0.1.2
neo4j>=5.18.0

# Memory & Storage
zep-python>=1.4.2
supabase>=2.3.0
PyJWT>=2.10.0
